+++ pyngrok/__init__.py
@@ -1,3 +1,3 @@
 __copyright__ = "Copyright (c) 2018-2025 Alex Laird"
 __license__ = "MIT"
-__version__ = "8.0.0"
+__version__ = "8.1.0"